+++ app/Sources/PiDCode/Models/WorkbenchModels.swift
@@ -11,6 +11,7 @@ enum SettingsPage: String, CaseIterable, Equatable, Sendable {
     case appearance
     case workbench
     case archivedSessions
+    case permissions
     case about
 }